haik8s/backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
sqlmodel>=0.0.14
aiosqlite>=0.20.0
python-jose[cryptography]>=3.3.0
httpx>=0.25.0
kubernetes==25.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
